refactor(ui): split account-service.tsx into focused sub-components #164
Extract 914-line monolith into 5 focused files under service-orders/:
- service-order-kanban.tsx (KanbanCard, KanbanColumn, badges, avatars)
- service-order-calendar.tsx (CalendarView + calendar utilities)
- create-order-modal.tsx (CreateOrderModal form)
- service-order-filter-bar.tsx (FilterBar + export config)
Main account-service.tsx reduced to 135 lines (composition only).
